requirements.txt
requests
openai
pydub
SpeechRecognition
pyaudio
pyserial
python-vlc